requirements.txt
click
Faker
kafka-python
requests